ISSUE #4 変更:  SearchScreenFragment で定義されていた CustomAdapter クラスを別ファイルに移動しました
