+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/SearchScreenFragment.kt
@@ -4,11 +4,8 @@
 package jp.co.yumemi.android.code_check
 
 import android.os.Bundle
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
 import android.view.inputmethod.EditorInfo
-import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.lifecycleScope
 import androidx.navigation.fragment.findNavController
@@ -62,43 +59,5 @@ class SearchScreenFragment : Fragment(R.layout.fragment_search_screen) {
         val action = SearchScreenFragmentDirections
             .actionNavigateToFragmentSearchResultsDetailScreen(githubRepository = item)
         findNavController().navigate(action)
-    }
-}
-
-val diffCallback = object : DiffUtil.ItemCallback<GithubRepository>() {
-    override fun areItemsTheSame(oldItem: GithubRepository, newItem: GithubRepository): Boolean {
-        return oldItem.name == newItem.name
-    }
-
-    override fun areContentsTheSame(oldItem: GithubRepository, newItem: GithubRepository): Boolean {
-        return oldItem == newItem
-    }
-
-}
-
-class CustomAdapter(
-    private val itemClickListener: OnItemClickListener,
-) : ListAdapter<GithubRepository, CustomAdapter.ViewHolder>(diffCallback) {
-
-    class ViewHolder(view: View) : RecyclerView.ViewHolder(view)
-
-    interface OnItemClickListener {
-        fun itemClick(item: GithubRepository)
-    }
-
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
-        val view = LayoutInflater.from(parent.context)
-            .inflate(R.layout.layout_item, parent, false)
-        return ViewHolder(view)
-    }
-
-    override fun onBindViewHolder(holder: ViewHolder, position: Int) {
-        val item = getItem(position)
-        (holder.itemView.findViewById<View>(R.id.repository_name_view) as TextView).text =
-            item.name
-
-        holder.itemView.setOnClickListener {
-            itemClickListener.itemClick(item)
-        }
     }
 }
